Check file names

Checks the file names according to ADV standard. File names should be in the following schema: prefix_utmzone_minx_miny_tilesize_region_year.laz

Usage

check_names(
  path,
  prefix = "3dm",
  zone = 32,
  region = NULL,
  year = NULL,
  copc = FALSE,
  full.names = FALSE,
  verbose = FALSE
)

Arguments

path

A path to a directory which contains las/laz files

prefix

3 letter character. Naming prefix (defaults to "3dm")

zone

2 digits integer. UTM zone (defaults to 32)

region

2 letter character. (optional) federal state abbreviation. It will be fetched automatically if not defined (default).

year

(optional) either an acquisition year (YYYY) or a proper csv file where to read the year. If not provided (default) the year will be extracted from the files. It will be the acquisition date if points contain datetime in GPStime format, otherwise it will get the year from the file header, which is the processing date by definition.

copc

Whether the file is expected to be a Cloud Optimized Point Cloud (.copc.laz)

full.names

Whether to return the full file paths or just the filenames (default) Whether to return the full file path or just the file names (default)

Value

A data.frame with attributes name_is, name_should, correct

Details

(e.g. 3dm_32_547_5724_1_ni_2024.laz)

Examples

f <- system.file("extdata", package = "managelidar")
check_names(f)
#> Error in loadNamespace(x): there is no package called ‘lasR’