Get data in a standardized format

This function copies data from one folder to another folder, while ensuring certain data formatting and folder structure. CRS is set, points are sorted, files are compressed, files are renamed according to ADV standard, files are ordered in folders by acquisition date and campaign, a VPC is created and files are spatially indexed.

Usage

get_data(origin, destination, campaign, origin_recurse = FALSE)

Arguments

origin

The path to a directory which contains (unprocessed) las/laz files

destination

The directory under which the processed files are copied and subfolders (year/campaign) are created

campaign

(character) Name of the project or campaign of data acquisition.

origin_recurse

boolean. Should files in subfolder be included?

Value

A structured copy of input lidar data

Examples

if (FALSE) { # \dontrun{
f <- system.file("extdata", package = "managelidar")
get_data(f, tempdir(), "landesbefliegung")
} # }