Get acquisition datetime from point cloud

get_datetime() derives the acquisition datetime for LAS files.

Usage

get_datetime(
  path,
  full.names = FALSE,
  verbose = FALSE,
  from_csv = NULL,
  return_referenceyear = FALSE
)

Arguments

path

The path to a file (.las/.laz/.copc), to a directory which contains these files, or to a virtual point cloud (.vpc) referencing these files.

full.names

Whether to return the full file paths or just the filenames (default) Whether to return the full file path or just the file name (default)

from_csv

When NULL (default) the datetime will be extracted from internal LAS data, if it points to a valid csv file the datetime will be read from this file.

return_referenceyear

If TRUE (FALSE is default) and a valid csv file is provided the reference year will be read from file instead of the date (e.g. referenceyear might be 2015 if data was acquired in december 2014).

Value

A data.frame with attributes filename, datetime_min and datetime_max

Details

This function derives acquistion dates for LAS files either from inheritted data or from an external csv file. Deriving dates from internal data requires the entire point cloud to be read! Thus it takes potentially a long processing time, especially if applied on entire folders or large files. Unfortunately it is only possible to derive exact acquisition date from newer (>= LAS 1.3) point cloud data, where acquisition date is encoded as adjusted standard GPS time. Otherwise, additional information on GPSweek is necessary. If from_csv points to a csv file with a date column and minx/miny in the same CRS than the LAS data, the date is extracted here. inst/extdata/acquisition_dates_lgln.csv contains dates for federal data in Lower Saxony.

Examples

f <- system.file("extdata", package = "managelidar")
get_datetime(f)
#>                                                                                                                                                                                                                                                                                                                                 
#>                            filename        datetime_min        datetime_max
#> 1 3dm_32_547_5724_1_ni_20240327.laz 2024-03-27 08:23:06 2024-03-27 08:29:25
#> 2 3dm_32_547_5725_1_ni_20240327.laz 2024-03-27 08:25:48 2024-03-27 08:32:28
#> 3 3dm_32_548_5724_1_ni_20240327.laz 2024-03-27 08:23:05 2024-03-27 08:36:29
#> 4 3dm_32_548_5725_1_ni_20240327.laz 2024-03-27 08:29:22 2024-03-27 08:40:51