Get intersecting LAS files

Get intersecting LAS files

Usage

get_intersection(path1, path2, mode = "intersects", full.names = FALSE)

Arguments

path1

path The path to a LAS file (.las/.laz/.copc), to a directory which contains LAS files, or to a Virtual Point Cloud (.vpc) referencing LAS files.

path2

The path to a LAS file (.las/.laz/.copc), to a directory which contains LAS files, or to a Virtual Point Cloud (.vpc) referencing LAS files.

mode

Either 'intersects' (default) or 'equals'.

full.names

Whether to return the full file paths or just the filenames (default)

Value

A spatial features data.frame with attributes filename, xmin, xmax, ymin, ymax of all files in path1 which intersect files in path2 ('intersects') or which have the same extent as files in path2 ('equals').

Examples

folder <- system.file("extdata", package = "managelidar")
file <- list.files(folder, full.names = T)[1]
get_intersection(folder, file)
#> Error in get_intersection(folder, file): could not find function "get_intersection"