Get the pulse penetration ratio of lasfiles

get_penetration() calculates the approximate pulse penetration ratio of lasfiles.

Usage

get_penetration(path, full.names = FALSE)

Arguments

path

A path to a laz file or a directory which contains laz files

full.names

Whether to return the full file path or just the file name (default)

Value

A dataframe with file, single, two, three, four, five, six, multiple

Details

For this function only the header from lasfiles is read. It calculates the ratio of pulses which have a single return only, which have two returns, three returns, four returns, five returns, six returns and for convenience multiple returns (two or more). Beware that this does not have to be exact especially for small files since pulses can be split at borders.

Examples

f <- system.file("extdata", package="managelidar")
get_penetration(f)
#>                                file single    two three  four  five   six
#> 1 3dm_32_547_5724_1_ni_20240327.laz  0.951  0.000 0.025 0.020 0.004 0.000
#> 2 3dm_32_547_5725_1_ni_20240327.laz  0.601 -0.160 0.122 0.315 0.120 0.001
#> 3 3dm_32_548_5724_1_ni_20240327.laz  0.417 -0.067 0.138 0.349 0.153 0.009
#> 4 3dm_32_548_5725_1_ni_20240327.laz  0.610 -0.105 0.147 0.234 0.111 0.004
#>   multiple
#> 1    0.049
#> 2    0.399
#> 3    0.583
#> 4    0.390